Big change, adding osruntime-bootstrap
-replaces code in mulitple places to launch the osruntime (chroot or
podman)
-osruntime-bootstrap is meant to be copied and modified with
run-specific var assignment, so it can be placed on host and run without
arguments. make_osruntime_boostrap_script() provides this functionality
-because of the disconnected nature in which osruntime-boostrap runs,
extensive error checking and logging is implemented, and as long as vars
for rickshaw_host and ssh_id and run_id are defined, the script should
be able to send errors back to the controller, and the endpoint script
[which executed the osruntime-bootstrap script] should be able to see
those errors and send abort for endpoint-deploy.
-osrumtime-boostrap's chroot execution needs work.  The env file, which
is generated by this script, needs to be sourced properly (probably
within the chroot).
-do_ssh is updated to not sudo for endpoint osp
-for osp endpoint, clear_all_servers implemented